CM4nowcast/requirements.txt
dask==2022.9.1
matplotlib==3.5.1
numba==0.55.1
numpy==1.21.0
pandas==1.4.1
properscoring==0.1
pytorch-lightning==1.5.10
torchvision==0.11.3
absl-py
boto3
javalang>=0.13.0
h5py>=2.10.0
yacs>=0.1.8
protobuf
unidiff
scipy
tqdm
regex
requests
jsonlines
contextvars
pyarrow>=3
transformers>=4.3.0
tensorboard
pandas
contextvars
dataclasses
pickle5
graphviz
networkx
fairscale>=0.3.0
fvcore>=0.1.5
pympler
einops>=0.3.0
timm
omegaconf
matplotlib
pysteps
pywavelets
pytorch_wavelets